[Streaming Generator] Handle out of order report when retry (#36069)
This PR handles edge cases for streaming generator tasks that have a nondeterministic number of return values. If the task is retried, we must notify the user with an error. In particular if the second task attempt returns a fewer number of return values than the first attempt, we:

    fail the remaining objects with the same exception
    include all known remaining objects (up to the max returned by the first attempt) when updating reference counts
    ignore all reports about task return values from the first attempt

---------

+++ src/ray/common/task/task_spec.h
@@ -254,7 +254,7 @@ class TaskSpecification : public MessageWrapper<rpc::TaskSpec> {
 
   size_t NumStreamingGeneratorReturns() const;
 
-  ObjectID StreamingGeneratorReturnId(uint32_t generator_index) const;
+  ObjectID StreamingGeneratorReturnId(size_t generator_index) const;
 
   void SetNumStreamingGeneratorReturns(uint64_t num_streaming_generator_returns);
